Add: Graph3D: improve error messages of setGraph3DFile functions
Existence of files passed to the args of the following functions
had not been checked, so they had dumped crash-reports
when the specified files were not found.
It had been very confusing for typical users.

* setGraph3DFile(string files[])
* setGraph3DFile(string files[], string[] modes)

Hence we added code for checking file-existence
and printing kind error messages any file is not found.