Add: Graph2D: improve error messages of setGraph2DFile function
Existence of files passed to the args of the following function
had not been checked, so they had dumped crash-reports
when the specified files were not found.

* setGraph2DFile(int graphID, string files[])

It had been very confusing for typical users.
Hence we added code for checking file-existence
and printing kind error messages
when any specified file is not found.